o1_utils::bitwise_operations

Trait BitwiseOps

pub trait BitwiseOps<Rhs = Self> {
    // Required methods
    fn bitwise_xor(input1: &Rhs, input: &Rhs) -> Rhs;
    fn bitwise_and(input1: &Rhs, input: &Rhs, bytes: usize) -> Rhs;
    fn bitwise_not(input: &Rhs, bits: Option<usize>) -> Rhs;
}

Bitwise operations

Required Methods

fn bitwise_xor(input1: &Rhs, input: &Rhs) -> Rhs

Bitwise XOR of two BigUint inputs

fn bitwise_and(input1: &Rhs, input: &Rhs, bytes: usize) -> Rhs

Conjunction of the bits of two BigUint inputs for a given number of bytes

fn bitwise_not(input: &Rhs, bits: Option<usize>) -> Rhs

Negate the bits of a Self input If it provides a larger desired bits than the input length then it takes the padded input of bits length. Otherwise it only takes the bits of the input.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.