o1_utils::field_helpers

Trait RandomField

pub trait RandomField<F> {
    // Required methods
    fn gen_field_with_bits(&mut self, bits: usize) -> F;
    fn gen(&mut self, input: Option<F>, bits: Option<usize>) -> F;
}

Helper to generate random field elements

Required Methods

fn gen_field_with_bits(&mut self, bits: usize) -> F

Generates a random field element of up to a given number of bits

fn gen(&mut self, input: Option<F>, bits: Option<usize>) -> F

Initialize a random input with a random value of given length