Crate o1_utils

A collection of utility functions and constants that can be reused from multiple projects

Re-exports

pub use dense_polynomial::ExtendedDensePolynomial;

pub use evaluations::ExtendedEvaluations;

pub use field_helpers::FieldHelpers;

pub use foreign_field::ForeignElement;

pub use foreign_field::LIMB_COUNT;

Modules

adjacent_pairs

This module hosts the AdjacentPairs type, which can be used to list all the adjacent pairs of a list. For example, if you have a list of integers [1, 2, 3], you can use it to obtain the list of tuples [(1, 2), (2, 3)].

chunked_polynomial

This module contains a type ChunkedPolynomial, and a number of helper methods to deal with chunked polynomials. Polynomials that cut in several polynomials of the same length.

dense_polynomial

This adds a few utility functions for the [DensePolynomial] arkworks type.

evaluations

This adds a few utility functions for the [Evaluations] arkworks type.

field_helpers

Useful helper methods to extend [ark_ff::Field].

foreign_field

Describes helpers for foreign field arithmetics

hasher

This module provides the CryptoDigest trait, which provides a generic interface for hashing.

math

This modules implements some math helper functions.

serialization

This adds a few utility functions for serializing and deserializing arkworks types that implement [CanonicalSerialize] and [CanonicalDeserialize].