o1_utils::dense_polynomial

Trait ExtendedDensePolynomial

pub trait ExtendedDensePolynomial<F: Field> {
    // Required methods
    fn scale(&self, elm: F) -> Self;
    fn shiftr(&self, size: usize) -> Self;
    fn eval_polynomial(coeffs: &[F], x: F) -> F;
    fn to_chunked_polynomial(
        &self,
        num_chunks: usize,
        size: usize,
    ) -> ChunkedPolynomial<F>;
}

An extension for the [DensePolynomial] type.

Required Methods

fn scale(&self, elm: F) -> Self

This function “scales” (multiplies all the coefficients of) a polynomial with a scalar.

fn shiftr(&self, size: usize) -> Self

Shifts all the coefficients to the right.

fn eval_polynomial(coeffs: &[F], x: F) -> F

eval_polynomial(coeffs, x) evaluates a polynomial given its coefficients coeffs and a point x.

fn to_chunked_polynomial( &self, num_chunks: usize, size: usize, ) -> ChunkedPolynomial<F>

Convert a polynomial into chunks. Implementors must ensure that the result contains exactly num_chunks.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.