Trait o1_utils::dense_polynomial::ExtendedDensePolynomial

pub trait ExtendedDensePolynomial<F: Field> {
    fn scale(&self, elm: F) -> Self;

    fn shiftr(&self, size: usize) -> Self;

    fn eval_polynomial(coeffs: &[F], x: F) -> F;

    fn eval(&self, elm: F, size: usize) -> Vec<F>;

    fn chunk_polynomial(&self, zeta_n: F, n: usize) -> Self;
}

An extension for the [DensePolynomial] type.

Required methods

fn scale(&self, elm: F) -> Self

This function “scales” (multiplies all the coefficients of) a polynomial with a scalar.

fn shiftr(&self, size: usize) -> Self

Shifts all the coefficients to the right.

fn eval_polynomial(coeffs: &[F], x: F) -> F

eval_polynomial(coeffs, x) evaluates a polynomial given its coefficients coeffs and a point x.

fn eval(&self, elm: F, size: usize) -> Vec<F>

This function evaluates polynomial in chunks.

fn chunk_polynomial(&self, zeta_n: F, n: usize) -> Self

Multiplies the chunks of a polynomial with powers of zeta^n to make it of degree n-1. For example, if a polynomial can be written f = f0 + x^n f1 + x^2n f2 (where f0, f1, f2 are of degree n-1), then this function returns the new semi-evaluated f'(x) = f0(x) + zeta^n f1(x) + zeta^2n f2(x).