o1_utils

Module field_helpers

Useful helper methods to extend [ark_ff::Field].

Enums

FieldHelpersError

Field helpers error

Traits

BigUintFieldHelpers

Field element wrapper for BigUint

FieldHelpers

Field element helpers Unless otherwise stated everything is in little-endian byte order.

RandomField

Helper to generate random field elements

Two

Helper to obtain two

Functions

i32_to_field

Converts an i32 into a [Field]

inner_prod

COmpute the inner product of two slices of field elements.

pows

pows(d, x) returns a vector containing the first d powers of the field element x (from 1 to x^(d-1)).

product

Returns the product of all the field elements belonging to an iterator.

Type Aliases

Result

Result alias using FieldHelpersError