o1_utils::foreign_field

Struct ForeignElement

pub struct ForeignElement<F: Field, const B: usize, const N: usize> {
    pub limbs: [F; N],
    /* private fields */
}

Represents a foreign field element Represents a foreign field element

Fields
limbs: [F; N]

limbs in little endian order

Implementations

impl<F: Field, const B: usize, const N: usize> ForeignElement<F, B, N>

pub fn new(limbs: [F; N]) -> Self

Creates a new foreign element from an array containing N limbs

pub fn zero() -> Self

Creates a new foreign element representing the value zero

pub fn from_biguint(big: BigUint) -> Self

Initializes a new foreign element from a big unsigned integer Panics if the BigUint is too large to fit in the N limbs

pub fn neg(&self, modulus: &BigUint) -> Self

Initializes a new foreign element from an absolute BigUint but the equivalent foreign element obtained corresponds to the negated input. It first converts the input big element to a big integer modulo the foreign field modulus, and then computes the negation of the result.

pub fn from_be(bytes: &[u8]) -> Self

Initializes a new foreign element from a set of bytes in big endian

pub fn to_biguint(&self) -> BigUint

Obtains the big integer representation of the foreign field element

impl<F: PrimeField, const B: usize, const N: usize> ForeignElement<F, B, N>

pub fn from_field(field: F) -> Self

Initializes a new foreign element from an element in the native field